Add some skeletal functionality to the Sidebar component
